move doesNotWantSpace to helpers
