+++ packages/language-support/src/formatting/formattingHelpers.ts
@@ -211,3 +211,14 @@ export function verifyGroupSizes(buffers: Chunk[][]) {
     }
   }
 }
+
+const openingCharacters = [CypherCmdLexer.LPAREN, CypherCmdLexer.LBRACKET];
+
+export function doesNotWantSpace(chunk: Chunk, nextChunk: Chunk): boolean {
+  return (
+    nextChunk?.type !== 'COMMENT' &&
+    chunk.type === 'REGULAR' &&
+    (chunk.noSpace ||
+      (chunk.node && openingCharacters.includes(chunk.node.symbol.type)))
+  );
+}
